requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.0
fonttools==4.47.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
kiwisolver==1.4.5
mangum==0.17.0
matplotlib==3.8.2
numpy==1.26.3
openai==1.9.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pluggy==1.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.4
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
